Parlay_Pete_Final/requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
python-dotenv==1.0.1
aiosqlite==0.20.0
